+++ backend/infra/rest/publicAuth.go
@@ -11,13 +11,11 @@ import (
 	"github.com/eclipse-disuko/disuko/conf"
 	"github.com/eclipse-disuko/disuko/domain/project"
 	"github.com/eclipse-disuko/disuko/domain/publicauth"
-	"github.com/eclipse-disuko/disuko/domain/user"
 	"github.com/eclipse-disuko/disuko/helper/exception"
 	"github.com/eclipse-disuko/disuko/helper/message"
 	"github.com/eclipse-disuko/disuko/helper/validation"
 	project2 "github.com/eclipse-disuko/disuko/infra/repository/project"
 	projectRepo "github.com/eclipse-disuko/disuko/infra/repository/project"
-	userRepo "github.com/eclipse-disuko/disuko/infra/repository/user"
 	"github.com/eclipse-disuko/disuko/logy"
 	"github.com/go-chi/render"
 	"github.com/golang-jwt/jwt/v4"
@@ -294,42 +292,4 @@ func projectTokenAuth(rs *logy.RequestSession, repo projectRepo.IProjectReposito
 		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid disco token"), "Project token not found or expired")
 	}
 	return parentToken
-}
-
-func patAuth(rs *logy.RequestSession, pr *project.Project, prRepo projectRepo.IProjectRepository, userRepo userRepo.IUsersRepository, tokenStr string) string {
-	token, err := jwt.ParseWithClaims(tokenStr, &user.UserTokenClaims{}, func(token *jwt.Token) (any, error) {
-		return []byte(conf.Config.Auth.UserTokenSigningKey), nil
-	})
-	if err != nil {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), err.Error())
-	}
-	claims, ok := token.Claims.(*user.UserTokenClaims)
-	if !ok {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), "Unexpected claims")
-	}
-	user := userRepo.FindByKey(rs, claims.UserKey, false)
-	if user == nil {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), "Unexpected claims")
-	}
-	ut := user.Token(claims.TokenKey)
-	if ut == nil {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), "Unexpected claims")
-	}
-	if ut.Expired() {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), "Unexpected claims")
-	}
-	if pr.GetMember(user.User) != nil {
-		return user.TokenOrigin(ut)
-	}
-	if !pr.HasParent() {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), "Project access denied")
-	}
-	parent := prRepo.FindByKey(rs, pr.Parent, false)
-	if parent == nil {
-		exception.ThrowExceptionServerMessage(message.GetI18N(message.ErrorDbNotFound), "Parent project not found")
-	}
-	if pr.GetMember(user.User) == nil {
-		exception.ThrowExceptionSendDeniedResponseRaw(message.GetI18N(message.DiscoTokenUnauthorized, "Invalid PAT"), "Project access denied")
-	}
-	return user.TokenOrigin(ut)
 }
